feat: coupon listing facet
