+++ packages/ats/contracts/contracts/facets/layer_2/coupon/CouponFacetBase.sol
@@ -7,25 +7,27 @@ import { IStaticFunctionSelectors } from "../../../infrastructure/proxy/IStaticF
 
 abstract contract CouponFacetBase is Coupon, IStaticFunctionSelectors {
     function getStaticFunctionSelectors() external pure override returns (bytes4[] memory staticFunctionSelectors_) {
-        staticFunctionSelectors_ = new bytes4[](12);
-        uint256 selectorIndex;
-        staticFunctionSelectors_[selectorIndex++] = this.setCoupon.selector;
-        staticFunctionSelectors_[selectorIndex++] = this.cancelCoupon.selector;
-        staticFunctionSelectors_[selectorIndex++] = this.getCoupon.selector;
-        staticFunctionSelectors_[selectorIndex++] = this.getCouponFor.selector;
-        staticFunctionSelectors_[selectorIndex++] = this.getCouponsFor.selector;
-        staticFunctionSelectors_[selectorIndex++] = this.getCouponAmountFor.selector;
-        staticFunctionSelectors_[selectorIndex++] = this.getCouponCount.selector;
-        staticFunctionSelectors_[selectorIndex++] = this.getCouponHolders.selector;
-        staticFunctionSelectors_[selectorIndex++] = this.getTotalCouponHolders.selector;
-        staticFunctionSelectors_[selectorIndex++] = this.getCouponFromOrderedListAt.selector;
-        staticFunctionSelectors_[selectorIndex++] = this.getCouponsOrderedList.selector;
-        staticFunctionSelectors_[selectorIndex++] = this.getCouponsOrderedListTotal.selector;
+        uint256 selectorIndex = 9;
+        staticFunctionSelectors_ = new bytes4[](selectorIndex);
+        unchecked {
+            staticFunctionSelectors_[--selectorIndex] = this.getTotalCouponHolders.selector;
+            staticFunctionSelectors_[--selectorIndex] = this.getCouponHolders.selector;
+            staticFunctionSelectors_[--selectorIndex] = this.getCouponCount.selector;
+            staticFunctionSelectors_[--selectorIndex] = this.getCouponAmountFor.selector;
+            staticFunctionSelectors_[--selectorIndex] = this.getCouponsFor.selector;
+            staticFunctionSelectors_[--selectorIndex] = this.getCouponFor.selector;
+            staticFunctionSelectors_[--selectorIndex] = this.getCoupon.selector;
+            staticFunctionSelectors_[--selectorIndex] = this.cancelCoupon.selector;
+            staticFunctionSelectors_[--selectorIndex] = this.setCoupon.selector;
+        }
     }
 
+    /// @inheritdoc IStaticFunctionSelectors
     function getStaticInterfaceIds() external pure override returns (bytes4[] memory staticInterfaceIds_) {
-        staticInterfaceIds_ = new bytes4[](1);
-        uint256 selectorsIndex;
-        staticInterfaceIds_[selectorsIndex++] = type(ICoupon).interfaceId;
+        uint256 selectorsIndex = 1;
+        staticInterfaceIds_ = new bytes4[](selectorsIndex);
+        unchecked {
+            staticInterfaceIds_[--selectorsIndex] = type(ICoupon).interfaceId;
+        }
     }
 }
