+++ packages/ats/contracts/scripts/domain/bondFixedRate/createConfiguration.ts
@@ -109,7 +109,7 @@ const BOND_FIXED_RATE_FACETS = [
   "ProtectedPartitionsFacet",
   "ScheduledBalanceAdjustmentsFacet",
   "ScheduledCrossOrderedTasksFacet",
-  "ScheduledCouponListingFacet",
+  "CouponListingFacet",
   "SsiManagementFacet",
 
   // Interest Rate (rate-specific)
