+++ packages/ats/contracts/scripts/domain/bondKpiLinkedRate/createConfiguration.ts
@@ -106,7 +106,7 @@ const BOND_KPI_LINKED_RATE_FACETS = [
   "ProtectedPartitionsFacet",
   "ScheduledBalanceAdjustmentsFacet",
   "ScheduledCrossOrderedTasksKpiLinkedRateFacet", // rate-specific: _onCouponListed override
-  "ScheduledCouponListingFacet",
+  "CouponListingFacet",
   "SsiManagementFacet",
   "TransferAndLockFacet",
