+++ packages/ats/contracts/scripts/domain/loanPortfolio/createConfiguration.ts
@@ -108,7 +108,7 @@ const LOANS_PORTFOLIO_FACETS = [
   // Scheduled Tasks
   "ScheduledBalanceAdjustmentsFacet",
   "ScheduledCrossOrderedTasksFacet",
-  "ScheduledCouponListingFacet",
+  "CouponListingFacet",
 
   // External Management
   "ExternalPauseManagementFacet",
