+++ packages/ats/contracts/test/contracts/integration/couponListing/couponListing.test.ts
@@ -1,0 +1,305 @@
+// SPDX-License-Identifier: Apache-2.0
+
+import { expect } from "chai";
+import { ethers } from "hardhat";
+import { HardhatEthersSigner } from "@nomicfoundation/hardhat-ethers/signers.js";
+import { type ResolverProxy, type IAsset } from "@contract-types";
+import { deployBondKpiLinkedRateTokenFixture, getDltTimestamp } from "@test";
+import { loadFixture } from "@nomicfoundation/hardhat-network-helpers";
+import { ATS_ROLES, TIME_PERIODS_S } from "@scripts";
+
+describe("CouponListing Tests", () => {
+  let diamond: ResolverProxy;
+  let signer_A: HardhatEthersSigner;
+
+  let asset: IAsset;
+
+  let startingDate = 0;
+  let maturityDate = 0;
+
+  async function deploySecurityFixture() {
+    const currentTimestamp = await getDltTimestamp();
+    startingDate = currentTimestamp + TIME_PERIODS_S.DAY;
+    maturityDate = startingDate + TIME_PERIODS_S.YEAR;
+
+    const base = await deployBondKpiLinkedRateTokenFixture({
+      bondDataParams: {
+        securityData: {
+          internalKycActivated: true,
+        },
+        bondDetails: {
+          startingDate,
+          maturityDate,
+        },
+      },
+    });
+
+    diamond = base.diamond;
+    signer_A = base.deployer;
+
+    asset = await ethers.getContractAt("IAsset", diamond.target);
+
+    await asset.grantRole(ATS_ROLES.CORPORATE_ACTION_ROLE, signer_A.address);
+  }
+
+  beforeEach(async () => {
+    await loadFixture(deploySecurityFixture);
+  });
+
+  // ─── getCouponsOrderedList / getCouponFromOrderedListAt / getCouponsOrderedListTotal ───
+
+  it("GIVEN multiple coupons WHEN triggerScheduledCrossOrderedTasks is called after fixingDate THEN coupons are added to ordered list", async () => {
+    const kpiLinkedRateBase = await deployBondKpiLinkedRateTokenFixture({
+      bondDataParams: {
+        securityData: {
+          isMultiPartition: false,
+        },
+        bondDetails: {
+          startingDate,
+          maturityDate,
+        },
+      },
+    });
+
+    const kpiDiamond = kpiLinkedRateBase.diamond;
+    const couponKpiLinkedRateFacet = await ethers.getContractAt(
+      "CouponKpiLinkedRateFacetTimeTravel",
+      kpiDiamond.target,
+      signer_A,
+    );
+    const accessControlKpi = await ethers.getContractAt("AccessControl", kpiDiamond.target, signer_A);
+    const scheduledTasksKpi = await ethers.getContractAt("ScheduledCrossOrderedTasks", kpiDiamond.target, signer_A);
+    const timeTravelKpi = await ethers.getContractAt("TimeTravelFacet", kpiDiamond.target, signer_A);
+
+    await accessControlKpi.connect(signer_A).grantRole(ATS_ROLES.CORPORATE_ACTION_ROLE, signer_A.address);
+
+    const timestamp = await getDltTimestamp();
+
+    const coupon1 = {
+      recordDate: (timestamp + TIME_PERIODS_S.DAY).toString(),
+      executionDate: (timestamp + TIME_PERIODS_S.DAY * 2).toString(),
+      rate: 0,
+      rateDecimals: 0,
+      startDate: timestamp.toString(),
+      endDate: (timestamp + TIME_PERIODS_S.DAY).toString(),
+      fixingDate: (timestamp + TIME_PERIODS_S.DAY).toString(),
+      rateStatus: 0,
+    };
+
+    const coupon2 = {
+      recordDate: (timestamp + TIME_PERIODS_S.DAY * 2).toString(),
+      executionDate: (timestamp + TIME_PERIODS_S.DAY * 3).toString(),
+      rate: 0,
+      rateDecimals: 0,
+      startDate: (timestamp + TIME_PERIODS_S.DAY).toString(),
+      endDate: (timestamp + TIME_PERIODS_S.DAY * 2).toString(),
+      fixingDate: (timestamp + TIME_PERIODS_S.DAY * 2).toString(),
+      rateStatus: 0,
+    };
+
+    await expect(couponKpiLinkedRateFacet.connect(signer_A).setCoupon(coupon1)).to.emit(
+      couponKpiLinkedRateFacet,
+      "CouponSet",
+    );
+
+    await expect(await couponKpiLinkedRateFacet.connect(signer_A).setCoupon(coupon2)).to.emit(
+      couponKpiLinkedRateFacet,
+      "CouponSet",
+    );
+
+    let orderedList = await couponKpiLinkedRateFacet.getCouponsOrderedList(0, 10);
+    expect(orderedList).to.be.an("array").with.lengthOf(0);
+
+    await timeTravelKpi.changeSystemTimestamp(timestamp + TIME_PERIODS_S.DAY + 1);
+
+    await scheduledTasksKpi.connect(signer_A).triggerScheduledCrossOrderedTasks(100);
+
+    orderedList = await couponKpiLinkedRateFacet.getCouponsOrderedList(0, 10);
+    expect(orderedList).to.be.an("array").with.lengthOf(1);
+    expect(orderedList[0]).to.equal(1); // couponId 1
+
+    await timeTravelKpi.changeSystemTimestamp(timestamp + TIME_PERIODS_S.DAY * 2 + 1);
+
+    await scheduledTasksKpi.connect(signer_A).triggerScheduledCrossOrderedTasks(100);
+
+    orderedList = await couponKpiLinkedRateFacet.getCouponsOrderedList(0, 10);
+    expect(orderedList).to.be.an("array").with.lengthOf(2);
+    expect(orderedList[0]).to.equal(1); // couponId 1
+    expect(orderedList[1]).to.equal(2); // couponId 2
+
+    // Verify getCouponFromOrderedListAt works correctly
+    const couponIdAtPos0 = await couponKpiLinkedRateFacet.getCouponFromOrderedListAt(0);
+    const couponIdAtPos1 = await couponKpiLinkedRateFacet.getCouponFromOrderedListAt(1);
+    expect(couponIdAtPos0).to.equal(1);
+    expect(couponIdAtPos1).to.equal(2);
+
+    // Verify getCouponsOrderedListTotal
+    const totalCouponsInOrderedList = await couponKpiLinkedRateFacet.getCouponsOrderedListTotal();
+    expect(totalCouponsInOrderedList).to.equal(2);
+  });
+
+  it("GIVEN deprecated coupons in bond storage AND new coupons in coupon storage WHEN getCouponFromOrderedListAt is called THEN it correctly routes to both storages with offset", async () => {
+    const kpiLinkedRateBase = await deployBondKpiLinkedRateTokenFixture({
+      bondDataParams: {
+        securityData: {
+          isMultiPartition: false,
+        },
+        bondDetails: {
+          startingDate,
+          maturityDate,
+        },
+      },
+    });
+
+    const kpiDiamond = kpiLinkedRateBase.diamond;
+    const couponKpiLinkedRateFacet = await ethers.getContractAt(
+      "CouponKpiLinkedRateFacetTimeTravel",
+      kpiDiamond.target,
+      signer_A,
+    );
+    const accessControlKpi = await ethers.getContractAt("AccessControl", kpiDiamond.target, signer_A);
+    const scheduledTasksKpi = await ethers.getContractAt("ScheduledCrossOrderedTasks", kpiDiamond.target, signer_A);
+    const timeTravelKpi = await ethers.getContractAt("TimeTravelFacet", kpiDiamond.target, signer_A);
+
+    await accessControlKpi.connect(signer_A).grantRole(ATS_ROLES.CORPORATE_ACTION_ROLE, signer_A.address);
+
+    const timestamp = await getDltTimestamp();
+
+    const coupon1 = {
+      recordDate: (timestamp + TIME_PERIODS_S.DAY).toString(),
+      executionDate: (timestamp + TIME_PERIODS_S.DAY * 2).toString(),
+      rate: 0,
+      rateDecimals: 0,
+      startDate: timestamp.toString(),
+      endDate: (timestamp + TIME_PERIODS_S.DAY).toString(),
+      fixingDate: (timestamp + TIME_PERIODS_S.DAY).toString(),
+      rateStatus: 0,
+    };
+
+    await expect(couponKpiLinkedRateFacet.connect(signer_A).setCoupon(coupon1)).to.emit(
+      couponKpiLinkedRateFacet,
+      "CouponSet",
+    );
+
+    // Trigger scheduled tasks to move coupon1 to ordered list
+    await timeTravelKpi.changeSystemTimestamp(timestamp + TIME_PERIODS_S.DAY + 1);
+    await scheduledTasksKpi.connect(signer_A).triggerScheduledCrossOrderedTasks(100);
+
+    // Add deprecated coupons to bond storage (simulating legacy data from before refactor)
+    // This must be done AFTER triggers complete to avoid errors when triggers try to process fake IDs
+    await timeTravelKpi.testOnlyAddDeprecatedCoupon(100);
+
+    // 1 deprecated (in bond storage) + 1 new (in coupon storage)
+    const totalCoupons = await couponKpiLinkedRateFacet.getCouponsOrderedListTotal();
+    expect(totalCoupons).to.equal(2);
+
+    const pos0 = await couponKpiLinkedRateFacet.getCouponFromOrderedListAt(0);
+    expect(pos0).to.equal(100); // from deprecated bond storage
+
+    const pos3 = await couponKpiLinkedRateFacet.getCouponFromOrderedListAt(1);
+    expect(pos3).to.equal(1); // first new coupon
+
+    const fullList = await couponKpiLinkedRateFacet.getCouponsOrderedList(0, 10);
+    expect(fullList).to.have.lengthOf(2);
+    expect(fullList[0]).to.equal(100); // deprecated
+    expect(fullList[1]).to.equal(1); // new storage
+  });
+
+  it("GIVEN empty ordered list WHEN getCouponFromOrderedListAt with _pos >= getCouponsOrderedListTotalAdjustedAt THEN returns 0", async () => {
+    const couponIdAtPos0 = await asset.getCouponFromOrderedListAt(0);
+    expect(couponIdAtPos0).to.equal(0);
+
+    // Test position 1 - should return 0 because list is empty
+    const couponIdAtPos1 = await asset.getCouponFromOrderedListAt(1);
+    expect(couponIdAtPos1).to.equal(0);
+
+    // Test position 100 - should return 0 because list is empty
+    const couponIdAtPos100 = await asset.getCouponFromOrderedListAt(100);
+    expect(couponIdAtPos100).to.equal(0);
+  });
+
+  // ─── scheduledCouponListingCount ───
+
+  describe("scheduledCouponListingCount", () => {
+    it("GIVEN no scheduled coupons WHEN scheduledCouponListingCount THEN returns 0", async () => {
+      const count = await asset.scheduledCouponListingCount();
+      expect(count).to.equal(0);
+    });
+
+    it("GIVEN scheduled coupons WHEN scheduledCouponListingCount THEN returns correct count", async () => {
+      for (let i = 0; i < 3; i++) {
+        const fixingDate = startingDate + TIME_PERIODS_S.MONTH * (i + 1);
+        const executionDate = fixingDate + TIME_PERIODS_S.WEEK;
+
+        await asset.setCoupon({
+          recordDate: fixingDate.toString(),
+          executionDate: executionDate.toString(),
+          rate: 0,
+          rateDecimals: 0,
+          startDate: (fixingDate - TIME_PERIODS_S.WEEK).toString(),
+          endDate: fixingDate.toString(),
+          fixingDate: fixingDate.toString(),
+          rateStatus: 0,
+        });
+      }
+
+      const count = await asset.scheduledCouponListingCount();
+      expect(count).to.equal(3);
+    });
+  });
+
+  // ─── getScheduledCouponListing ───
+
+  describe("getScheduledCouponListing", () => {
+    beforeEach(async () => {
+      for (let i = 0; i < 5; i++) {
+        const fixingDate = startingDate + TIME_PERIODS_S.MONTH * (i + 1);
+        const executionDate = fixingDate + TIME_PERIODS_S.WEEK;
+
+        await asset.setCoupon({
+          recordDate: fixingDate.toString(),
+          executionDate: executionDate.toString(),
+          rate: 0,
+          rateDecimals: 0,
+          startDate: (fixingDate - TIME_PERIODS_S.WEEK).toString(),
+          endDate: fixingDate.toString(),
+          fixingDate: fixingDate.toString(),
+          rateStatus: 0,
+        });
+      }
+    });
+
+    it("GIVEN scheduled coupons WHEN getScheduledCouponListing with page 0 and length 10 THEN returns all coupons", async () => {
+      const coupons = await asset.getScheduledCouponListing(0, 10);
+      expect(coupons.length).to.equal(5);
+    });
+
+    it("GIVEN scheduled coupons WHEN getScheduledCouponListing with page 0 and length 3 THEN returns first 3 coupons", async () => {
+      const coupons = await asset.getScheduledCouponListing(0, 3);
+      expect(coupons.length).to.equal(3);
+    });
+
+    it("GIVEN scheduled coupons WHEN getScheduledCouponListing with page 1 and length 3 THEN returns next 2 coupons", async () => {
+      const coupons = await asset.getScheduledCouponListing(1, 3);
+      expect(coupons.length).to.equal(2);
+    });
+
+    it("GIVEN scheduled coupons WHEN getScheduledCouponListing with page 2 and length 3 THEN returns empty array", async () => {
+      const coupons = await asset.getScheduledCouponListing(2, 3);
+      expect(coupons.length).to.equal(0);
+    });
+
+    it("GIVEN scheduled coupons WHEN getScheduledCouponListing THEN returns tasks with correct structure", async () => {
+      const coupons = await asset.getScheduledCouponListing(0, 1);
+      expect(coupons.length).to.equal(1);
+      const coupon = {
+        scheduledTimestamp: coupons[0].scheduledTimestamp,
+        data: coupons[0].data,
+      };
+      expect(coupon).to.have.property("scheduledTimestamp");
+      expect(coupon).to.have.property("data");
+      expect(coupon.scheduledTimestamp).to.be.gt(0);
+      expect(coupon.data).to.not.equal("0x");
+    });
+  });
+});
