+++ packages/ats/contracts/test/contracts/integration/layer_1/coupon/coupon.test.ts
@@ -868,174 +868,4 @@ describe("Coupon Tests", () => {
       "WrongTimestamp",
     );
   });
-
-  it("GIVEN multiple coupons WHEN triggerScheduledCrossOrderedTasks is called after fixingDate THEN coupons are added to ordered list", async () => {
-    const kpiLinkedRateBase = await deployBondKpiLinkedRateTokenFixture({
-      bondDataParams: {
-        securityData: {
-          isMultiPartition: false,
-        },
-        bondDetails: {
-          startingDate: startingDate,
-          maturityDate: maturityDate,
-        },
-      },
-    });
-
-    const kpiDiamond = kpiLinkedRateBase.diamond;
-    const couponKpiLinkedRateFacet = await ethers.getContractAt(
-      "CouponKpiLinkedRateFacetTimeTravel",
-      kpiDiamond.target,
-      signer_A,
-    );
-    const accessControlKpi = await ethers.getContractAt("AccessControl", kpiDiamond.target, signer_A);
-    const scheduledTasksKpi = await ethers.getContractAt("ScheduledCrossOrderedTasks", kpiDiamond.target, signer_A);
-    const timeTravelKpi = await ethers.getContractAt("TimeTravelFacet", kpiDiamond.target, signer_A);
-
-    await accessControlKpi.connect(signer_A).grantRole(ATS_ROLES.CORPORATE_ACTION_ROLE, signer_A.address);
-
-    const timestamp = await getDltTimestamp();
-
-    const coupon1 = {
-      recordDate: (timestamp + TIME_PERIODS_S.DAY).toString(),
-      executionDate: (timestamp + TIME_PERIODS_S.DAY * 2).toString(),
-      rate: 0,
-      rateDecimals: 0,
-      startDate: timestamp.toString(),
-      endDate: (timestamp + TIME_PERIODS_S.DAY).toString(),
-      fixingDate: (timestamp + TIME_PERIODS_S.DAY).toString(),
-      rateStatus: 0,
-    };
-
-    const coupon2 = {
-      recordDate: (timestamp + TIME_PERIODS_S.DAY * 2).toString(),
-      executionDate: (timestamp + TIME_PERIODS_S.DAY * 3).toString(),
-      rate: 0,
-      rateDecimals: 0,
-      startDate: (timestamp + TIME_PERIODS_S.DAY).toString(),
-      endDate: (timestamp + TIME_PERIODS_S.DAY * 2).toString(),
-      fixingDate: (timestamp + TIME_PERIODS_S.DAY * 2).toString(),
-      rateStatus: 0,
-    };
-
-    await expect(couponKpiLinkedRateFacet.connect(signer_A).setCoupon(coupon1)).to.emit(
-      couponKpiLinkedRateFacet,
-      "CouponSet",
-    );
-
-    await expect(await couponKpiLinkedRateFacet.connect(signer_A).setCoupon(coupon2)).to.emit(
-      couponKpiLinkedRateFacet,
-      "CouponSet",
-    );
-
-    let orderedList = await couponKpiLinkedRateFacet.getCouponsOrderedList(0, 10);
-    expect(orderedList).to.be.an("array").with.lengthOf(0);
-
-    await timeTravelKpi.changeSystemTimestamp(timestamp + TIME_PERIODS_S.DAY + 1);
-
-    await scheduledTasksKpi.connect(signer_A).triggerScheduledCrossOrderedTasks(100);
-
-    orderedList = await couponKpiLinkedRateFacet.getCouponsOrderedList(0, 10);
-    expect(orderedList).to.be.an("array").with.lengthOf(1);
-    expect(orderedList[0]).to.equal(1); // couponId 1
-
-    await timeTravelKpi.changeSystemTimestamp(timestamp + TIME_PERIODS_S.DAY * 2 + 1);
-
-    await scheduledTasksKpi.connect(signer_A).triggerScheduledCrossOrderedTasks(100);
-
-    orderedList = await couponKpiLinkedRateFacet.getCouponsOrderedList(0, 10);
-    expect(orderedList).to.be.an("array").with.lengthOf(2);
-    expect(orderedList[0]).to.equal(1); // couponId 1
-    expect(orderedList[1]).to.equal(2); // couponId 2
-
-    // Verify getCouponFromOrderedListAt works correctly
-    const couponIdAtPos0 = await couponKpiLinkedRateFacet.getCouponFromOrderedListAt(0);
-    const couponIdAtPos1 = await couponKpiLinkedRateFacet.getCouponFromOrderedListAt(1);
-    expect(couponIdAtPos0).to.equal(1);
-    expect(couponIdAtPos1).to.equal(2);
-
-    // Verify getCouponsOrderedListTotal
-    const totalCouponsInOrderedList = await couponKpiLinkedRateFacet.getCouponsOrderedListTotal();
-    expect(totalCouponsInOrderedList).to.equal(2);
-  });
-
-  it("GIVEN deprecated coupons in bond storage AND new coupons in coupon storage WHEN getCouponFromOrderedListAt is called THEN it correctly routes to both storages with offset", async () => {
-    const kpiLinkedRateBase = await deployBondKpiLinkedRateTokenFixture({
-      bondDataParams: {
-        securityData: {
-          isMultiPartition: false,
-        },
-        bondDetails: {
-          startingDate: startingDate,
-          maturityDate: maturityDate,
-        },
-      },
-    });
-
-    const kpiDiamond = kpiLinkedRateBase.diamond;
-    const couponKpiLinkedRateFacet = await ethers.getContractAt(
-      "CouponKpiLinkedRateFacetTimeTravel",
-      kpiDiamond.target,
-      signer_A,
-    );
-    const accessControlKpi = await ethers.getContractAt("AccessControl", kpiDiamond.target, signer_A);
-    const scheduledTasksKpi = await ethers.getContractAt("ScheduledCrossOrderedTasks", kpiDiamond.target, signer_A);
-    const timeTravelKpi = await ethers.getContractAt("TimeTravelFacet", kpiDiamond.target, signer_A);
-
-    await accessControlKpi.connect(signer_A).grantRole(ATS_ROLES.CORPORATE_ACTION_ROLE, signer_A.address);
-
-    const timestamp = await getDltTimestamp();
-
-    const coupon1 = {
-      recordDate: (timestamp + TIME_PERIODS_S.DAY).toString(),
-      executionDate: (timestamp + TIME_PERIODS_S.DAY * 2).toString(),
-      rate: 0,
-      rateDecimals: 0,
-      startDate: timestamp.toString(),
-      endDate: (timestamp + TIME_PERIODS_S.DAY).toString(),
-      fixingDate: (timestamp + TIME_PERIODS_S.DAY).toString(),
-      rateStatus: 0,
-    };
-
-    await expect(couponKpiLinkedRateFacet.connect(signer_A).setCoupon(coupon1)).to.emit(
-      couponKpiLinkedRateFacet,
-      "CouponSet",
-    );
-
-    // Trigger scheduled tasks to move coupon1 to ordered list
-    await timeTravelKpi.changeSystemTimestamp(timestamp + TIME_PERIODS_S.DAY + 1);
-    await scheduledTasksKpi.connect(signer_A).triggerScheduledCrossOrderedTasks(100);
-
-    // Add deprecated coupons to bond storage (simulating legacy data from before refactor)
-    // This must be done AFTER triggers complete to avoid errors when triggers try to process fake IDs
-    await timeTravelKpi.testOnlyAddDeprecatedCoupon(100);
-
-    // 1 deprecated (in bond storage) + 1 new (in coupon storage)
-    const totalCoupons = await couponKpiLinkedRateFacet.getCouponsOrderedListTotal();
-    expect(totalCoupons).to.equal(2);
-
-    const pos0 = await couponKpiLinkedRateFacet.getCouponFromOrderedListAt(0);
-    expect(pos0).to.equal(100); // from deprecated bond storage
-
-    const pos3 = await couponKpiLinkedRateFacet.getCouponFromOrderedListAt(1);
-    expect(pos3).to.equal(1); // first new coupon
-
-    const fullList = await couponKpiLinkedRateFacet.getCouponsOrderedList(0, 10);
-    expect(fullList).to.have.lengthOf(2);
-    expect(fullList[0]).to.equal(100); // deprecated
-    expect(fullList[1]).to.equal(1); // new storage
-  });
-
-  it("GIVEN empty ordered list WHEN getCouponFromOrderedListAt with _pos >= getCouponsOrderedListTotalAdjustedAt THEN returns 0", async () => {
-    const couponIdAtPos0 = await asset.getCouponFromOrderedListAt(0);
-    expect(couponIdAtPos0).to.equal(0);
-
-    // Test position 1 - should return 0 because list is empty
-    const couponIdAtPos1 = await asset.getCouponFromOrderedListAt(1);
-    expect(couponIdAtPos1).to.equal(0);
-
-    // Test position 100 - should return 0 because list is empty
-    const couponIdAtPos100 = await asset.getCouponFromOrderedListAt(100);
-    expect(couponIdAtPos100).to.equal(0);
-  });
 });
